// frontend/src/app/data/model/page-recipe.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Pageable } from './pageable';
import { Recipe } from './recipe';
import { Sort } from './sort';


export interface PageRecipe { 
    totalElements?: number;
    totalPages?: number;
    numberOfElements?: number;
    first?: boolean;
    last?: boolean;
    pageable?: Pageable;
    number?: number;
    size?: number;
    content?: Array<Recipe>;
    sort?: Sort;
    empty?: boolean;
}